pyproject.toml
[tool.poetry]
name = "ldraw-parts"
version = "0.1.0"
description = ""
authors = ["Matthieu Berthomé <matthieu@mmea.fr>"]

[tool.poetry.dependencies]
python = "^3.8.5"
pymklist = "^1.1.0"
requests = "^2.25.1"
requests-cache = "^0.5.2"
GitPython = "^3.1.12"
ordereddict = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
